rfp_semantic_pipeline/app/core/files.py
import hashlib
import json
from pathlib import Path
from typing import Any

import yaml


def ensure_dirs(paths: list[Path]) -> None:
    for path in paths:
        path.mkdir(parents=True, exist_ok=True)


def compute_sha256(content: bytes) -> str:
    return hashlib.sha256(content).hexdigest()


def dump_json(path: Path, payload: Any) -> None:
    path.parent.mkdir(parents=True, exist_ok=True)
    path.write_text(json.dumps(payload, ensure_ascii=False, indent=2), encoding="utf-8")


def load_taxonomy(path: Path) -> list[str]:
    data = yaml.safe_load(path.read_text(encoding="utf-8"))
    return data.get("categories", [])
