rfp_semantic_pipeline/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
streamlit==1.48.0
sqlmodel==0.0.24
PyMuPDF==1.26.3
pdfplumber==0.11.7
sentence-transformers==5.1.0
faiss-cpu==1.12.0
networkx==3.5
python-multipart==0.0.20
httpx==0.28.1
PyYAML==6.0.2
pandas==2.3.2
pytest==8.4.1
